GaussianMixtureAlignment

Documentation for GaussianMixtureAlignment.

GaussianMixtureAlignment.GaussianMixtureAlignment

GaussianMixtureAlignment.IsotropicGMM

GaussianMixtureAlignment.IsotropicGaussian

GaussianMixtureAlignment.IsotropicMultiGMM

GaussianMixtureAlignment.MultiPointSet

GaussianMixtureAlignment.Point

GaussianMixtureAlignment.PointSet

GaussianMixtureAlignment.UncertaintyRegion

GaussianMixtureAlignment.branchbound

GaussianMixtureAlignment.combine

GaussianMixtureAlignment.distance

GaussianMixtureAlignment.gauss_l2_bounds

GaussianMixtureAlignment.inertial_transforms

GaussianMixtureAlignment.local_align

GaussianMixtureAlignment.mass_matrix

GaussianMixtureAlignment.overlap

GaussianMixtureAlignment.overlap

GaussianMixtureAlignment.overlap

GaussianMixtureAlignment.overlap

GaussianMixtureAlignment.overlap

GaussianMixtureAlignment.rocs_align

GaussianMixtureAlignment.rot_branchbound

GaussianMixtureAlignment.subranges

GaussianMixtureAlignment.tanimoto

GaussianMixtureAlignment.tight_distance_bounds

GaussianMixtureAlignment.tivgmm

GaussianMixtureAlignment.tivpointset

GaussianMixtureAlignment.translation_limit

GaussianMixtureAlignment.trl_branchbound

GaussianMixtureAlignment.GaussianMixtureAlignment — Module

GaussianMixtureAlignment.jl

GaussianMixtureAlignment.jl is a package used to align Gaussian mixture models. In particular, it uses an implementation of the GOGMA algorithm (Campbell, 2016) to find globally optimal alignments of mixtures of isotropic (spherical) Gaussian distributions.

REPL help

? followed by an algorith or constructor name will print help to the terminal. See:

?IsotropicGaussian 

	?IsotropicGMM 

	?IsotropicMultiGMM 

	?gogma_align 

	?tiv_gogma_align 

	?rocs_align

GaussianMixtureAlignment.IsotropicGMM — Type

A collection of IsotropicGaussians, making up a Gaussian Mixture Model (GMM).

GaussianMixtureAlignment.IsotropicGaussian — Type

A structure that defines an isotropic Gaussian distribution with the location of the mean, μ, standard deviation σ, and scaling factor ϕ.

An IsotropicGaussian can also be assigned directions dirs which enforce a penalty for misalignment with the dirs of another IsotropicGaussian.

GaussianMixtureAlignment.IsotropicMultiGMM — Type

A collection of labeled IsotropicGMMs, to each be considered separately during an alignment procedure. That is, only alignment scores between IsotropicGMMs with the same key are considered when aligning two MultiGMMs.

GaussianMixtureAlignment.MultiPointSet — Type

A collection of labeled point sets, to each be considered separately during an alignment procedure. That is, only alignment scores between point sets with the same key are considered when aligning two MultiPointSets.

GaussianMixtureAlignment.Point — Type

A coordinate position and a weight, to be used as part of a point set.

GaussianMixtureAlignment.PointSet — Type

A point set made consisting of a matrix of coordinate positions with corresponding weights.

GaussianMixtureAlignment.UncertaintyRegion — Type

Describes an transformation uncertainty region centered at rotation R and translation T, with rotation and translation half-widths of σᵣ and σₜ respectively

GaussianMixtureAlignment.branchbound — Method

result = branchbound(x, y; nsplits=2, searchspace=nothing, 
                     rot=nothing, trl=nothing, blockfun=fullBlock, objfun=alignment_objective,
                     rtol=0.01, maxblocks=5e8, maxeva ls=Inf, maxstagnant=Inf, threads=false)

Finds the globally optimal rigid transform for alignment between two isotropic Gaussian mixtures, x and y, using the GOGMA algorithm.

Returns a GlobalAlignmentResult that contains the maximized overlap of the two GMMs (the upperbound on the objective function), a lower bound on the alignment objective function, an AffineMap which aligns x with y, and information about the number of evaluations during the alignment procedure.

GaussianMixtureAlignment.combine — Method

gmm = combine(gmmx::IsotropicGMM, gmmy::IsotropicGMM)
mgmm = combine(mgmmx::MultiGMM, mgmmy::MultiGMM)
gmm = combine(gmms::Union{AbstractVector{<:IsotropicGMM},AbstractVector{<:MultiGMM}})

Creates a new IsotropicGMM or MultiGMM by concatenating the vectors of IsotroicGaussians in the input GMMs.

GaussianMixtureAlignment.distance — Method

l2dist = distance(x, y)

Calculates the L2 distance between two GMMs made up of spherical Gaussian distributions.

GaussianMixtureAlignment.gauss_l2_bounds — Function

lowerbound, upperbound = gauss_l2_bounds(x::Union{IsotropicGaussian, AbstractGMM}, y::Union{IsotropicGaussian, AbstractGMM}, σᵣ, σₜ)
lowerbound, upperbound = gauss_l2_bounds(x, y, R::RotationVec, T::SVector{3}, σᵣ, σₜ)

Finds the bounds for overlap between two isotropic Gaussian distributions, two isotropic GMMs, or two sets of labeled isotropic GMMs for a particular region in 6-dimensional rigid rotation space, defined byR,T,σᵣandσₜ`.

R and T represent the rotation and translation, respectively, that are at the center of the uncertainty region. If they are not provided, the uncertainty region is assumed to be centered at the origin (i.e. x has already been transformed).

σᵣ and σₜ represent the sizes of the rotation and translation uncertainty regions.

See Campbell & Peterson, 2016

GaussianMixtureAlignment.inertial_transforms — Function

tforms = inertial_transforms(gmm)

Returns 10 transformations to put gmm in an inertial frame. That is, the mass matrix of the GMM is made diagonal, and the GMM center of mass is made the origin.

GaussianMixtureAlignment.local_align — Method

obj, pos = local_align(x, y, block, pσ=nothing, pϕ=nothing; R=nothing, T=nothing, maxevals=100)

Performs local alignment within the specified block using L-BFGS to minimize objective function objfun for the provided GMMs, x and y.

GaussianMixtureAlignment.mass_matrix — Function

m = second_moment(gmm, center, dim1, dim2)

Returns the second order moment of gmm

GaussianMixtureAlignment.overlap — Function

ovlp = overlap(x::AbstractSingleGMM, y::AbstractSingleGMM, xtform=identity)

Calculates the unnormalized overlap between two AbstractSingleGMM objects.

GaussianMixtureAlignment.overlap — Function

ovlp = overlap(x::AbstractMultiGMM, y::AbstractMultiGMM, xtform=identity)

Calculates the unnormalized overlap between two AbstractMultiGMM objects.

GaussianMixtureAlignment.overlap — Function

ovlp = overlap(x::IsotropicGaussian, y::IsotropicGaussian, xtform=identity)

Calculates the unnormalized overlap between two IsotropicGaussian objects.

GaussianMixtureAlignment.overlap — Method

ovlp = overlap(distsq, s, w, dirdot)

Calculates the unnormalized overlap between two Gaussian distributions with width s, weight w', and squared distancedistsq, and geometric scaling factordirdot`.

GaussianMixtureAlignment.overlap — Method

ovlp = overlap(dist, σx, σy, ϕx, ϕy, dirdot)

Calculates the unnormalized overlap between two Gaussian distributions with variances σx and σy, weights ϕx and ϕy, and means separated by distance dist, scaled by the dot product obtained from geometric constraints dirdot.

GaussianMixtureAlignment.rocs_align — Method

score, tform, nevals = rocs_align_gmms(gmmfixed, gmmmoving; maxevals=1000)

Finds the optimal alignment between the two supplied GMMs using steric multipoles, based on the ROCS alignment algorithm.

GaussianMixtureAlignment.rot_branchbound — Method

result = rot_gogma_align(x, y; kwargs...)

Finds the globally optimal rotation for alignment between two isotropic Gaussian mixtures, x and y, using the GOGMA algorithm, for a given rotation, rot. That is, only rigid translation is allowed.

For details about keyword arguments, see gogma_align().

GaussianMixtureAlignment.subranges — Function

sbrngs = subranges(ranges, nsplits)

Takes ranges, a nested tuple describing intervals for each dimension in rigid-transformation space defining a hypercube, and splits the hypercube into nsplits even components along each dimension. If the cube is N-dimensional, the number of returned sub-cubes will be nsplits^N.

GaussianMixtureAlignment.tanimoto — Method

tani = tanimoto(x, y)

Calculates the tanimoto distance based on Gaussian overlap between two GMMs.

GaussianMixtureAlignment.tight_distance_bounds — Method

lb, ub = tight_distance_bounds(x::SVector{3,<:Number}, y::SVector{3,<:Number}, σᵣ::Number, σₜ::Number)
lb, ub = tight_distance_bounds(x::SVector{3,<:Number}, y::SVector{3,<:Number}, R::RotationVec, T<:SVector{3}, σᵣ::Number, σₜ::Number)

Within an uncertainty region, find the bounds on distance between two points x and y.

See Campbell & Peterson, 2016

GaussianMixtureAlignment.tivgmm — Function

tgmm = tivgmm(gmm::IsotropicGMM, c=Inf)
tgmm = tivgmm(mgmm::MultiGMM, c=Inf)

Returns a new IsotropicGMM or MultiGMM containing up to c*length(gmm) translation invariant vectors (TIVs) connecting Gaussian means in gmm. TIVs are chosen to maximize length multiplied by the weights of the connected distributions.

See Li et. al. (2019) for a description of TIV construction.

GaussianMixtureAlignment.tivpointset — Function

tset = tivpointset(ps::PointSet, c=Inf)
tset = tivpointset(mps::MultiPointSet, c=Inf)

Returns a new PointSet or MultiPointSet containing up to c*length(gmm) translation invariant vectors (TIVs) connecting point coordinates in the input model. TIVs are chosen to maximize length multiplied by the weights of the connected distributions.

See Li et. al. (2019) for a description of TIV construction.

GaussianMixtureAlignment.translation_limit — Method

lim = translation_limit(gmmx, gmmy)

Computes the largest translation needed to ensure that the searchspace contains the best alignment transformation.

GaussianMixtureAlignment.trl_branchbound — Method

result = trl_gogma_align(x, y; kwargs...)

Finds the globally optimal translation for alignment between two isotropic Gaussian mixtures, x and y, using the GOGMA algorithm, for a given translation, trl. That is, only rigid rotation is allowed.

For details about keyword arguments, see gogma_align().